fix(workspace): break circular import between line-utils and search
Move findLineRange from line-utils.ts to search/bm25.ts since it
uses the tokenize function. This breaks the cycle where line-utils
imported from search and search imported from line-utils.
