+++ packages/core/src/workspace/search/search-engine.ts
@@ -8,8 +8,7 @@
 import type { SearchMode, LineRange } from '../../artifacts';
 import type { MastraVector, VectorFilter } from '../../vector';
 
-import { findLineRange } from '../line-utils';
-import { BM25Index, tokenize } from './bm25';
+import { BM25Index, tokenize, findLineRange } from './bm25';
 import type { BM25Config, TokenizeOptions } from './bm25';
 
 // Re-export SearchMode for consumers
